+++ src/Sekiban.Core/Command/ICommandContextWithoutGetState.cs
@@ -44,4 +44,5 @@ public interface ICommandContextWithoutGetState<TAggregatePayload> where TAggreg
     public Task<ResultBox<TOutput>> ExecuteQueryAsync<TOutput>(IQueryInput<TOutput> param) where TOutput : IQueryResponse;
 
     public ResultBox<UnitValue> AppendEvent(IEventPayloadApplicableTo<TAggregatePayload> eventPayload);
+    public ResultBox<UnitValue> AppendEvents(params IEventPayloadApplicableTo<TAggregatePayload>[] eventPayloads);
 }
